bitstring measurement and reformat
