+++ qujax/__init__.py
@@ -10,13 +10,13 @@ from qujax.circuit_tools import check_unitary
 from qujax.circuit_tools import check_circuit
 from qujax.circuit_tools import print_circuit
 
+from qujax.observable import statetensor_to_densitytensor
 from qujax.observable import statetensor_to_single_expectation
 from qujax.observable import densitytensor_to_single_expectation
 from qujax.observable import get_statetensor_to_expectation_func
 from qujax.observable import get_statetensor_to_sampled_expectation_func
 from qujax.observable import get_densitytensor_to_expectation_func
 from qujax.observable import get_densitytensor_to_sampled_expectation_func
-from qujax.observable import statetensor_to_densitytensor
 from qujax.observable import check_hermitian
 from qujax.observable import integers_to_bitstrings
 from qujax.observable import bitstrings_to_integers
